requirements.txt
Flask==3.0.0
Werkzeug==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.11.0
pytest==7.4.3
pytest-cov==4.1.0
gunicorn==21.2.0
